Commit moar
Kick is the most-working

+++ src/moody/commands/mod/clearwarns.js
@@ -7,13 +7,13 @@ const db = require('../../../models/warns')
 module.exports = class ClearWarnsCommand extends ModCommand {
     constructor(client) {
         let comprops = {
-            name: "removewarns",
+            name: "clearwarns",
             aliases: [
-                "clearwarns",
+                "removewarns",
                 "clrwarns"
             ],
-            group: "admin",
-            memberName: "removewarns",
+            group: "mod",
+            memberName: "clearwarns",
             description: "Clears all warns in server for user",
             guildOnly: true,
             clientPermissions: [
@@ -21,40 +21,73 @@ module.exports = class ClearWarnsCommand extends ModCommand {
             ],
             userPermissions: [
                 "KICK_MEMBERS"
+            ],
+            args: [
+                {
+                    key: "target",
+                    prompt: "User to clear warns for?",
+                    type: "member|user"
+                }
             ]
         }
         super(
             client,
             //@ts-ignore
-            {...comprops}
+            {...comprops},
+            {
+                description: ""
+            }
         )
     }
 
-    async action(client, message) {
+    async action(message, args) {
 
-        const user = this.inputData.loaded
+        let target = args.target
+        console.log(`Target In: ${target.displayName} (ID:${target.id})`)
+        console.log(JSON.stringify(target))
+        console.log(`ID:      ${target.id}`)
+        console.log(`?.ID:    ${target?.id}`)
+        console.log(`!?.ID:   ${!target?.id}`)
+        console.log(`!(?.ID): ${!(target?.id)}`)
 
-        if (!user) {
+        if (!(target?.id)) {
+            console.log(`No Target ID`)
+            target = await message.guild.members.find(
+                (user) =>
+                    (user?.id     === target) ||
+                    (user?.userID === target) ||
+                    (user?.name   === target) ||
+                    false
+            )
+            console.log(`Searched and got: ${target}`)
+            if(target?.user) {
+                target = target.user
+                console.log(`User: ${target}`)
+            }
+        }
+
+        if (!(target?.id)) {
             this.error = true
             this.props.description = this.errors.cantActionSelf
+            return
         }
 
         if(!(this.error)) {
             db.findOne({
                 guildID: message.guild.id,
-                user: user.id
+                user: target.id
             }, async (err, data) => {
                 if (err) throw err;
                 let props = { caption: { text: "Clear Warns" } }
                 if (data) {
                     await db.findOneAndDelete({
                         guildID: message.guild.id,
-                        user: user.id
+                        user: target.id
                     })
-                    props.description = `Cleared <@${user.id}>'s warns`
+                    props.description = `Cleared <@${target.id}>'s warns`
                 } else {
                     props.error = true
-                    props.description = `<@${user.id}> has no warns!`
+                    props.description = `<@${target.id}> has no warns!`
                 }
                 let embed = new VillainsEmbed(props)
                 // message.channel.send({ embeds: [embed] }) // discord.js v13
@@ -69,16 +102,16 @@ module.exports = class ClearWarnsCommand extends ModCommand {
         const baseArgs = []
         const varArgs = [
           "",
-          message.author.username,
-          message.author.id,
-          client.user.username,
-          "Wanrae"
+          { target: message.author.username },
+          { target: message.author.id },
+          { target: client.user.username },
+          { target: "Wanrae" }
         ]
 
         for(let added of varArgs) {
-            let args = baseArgs.concat([ ...added.split(" ") ])
+            let args = added
             dummy = new ClearWarnsCommand(client)
-            dummy.props.footer.msg = args.join(" | ")
+            dummy.props.footer.msg = typeof args === "object" && typeof args.join === "function" ? args.join(" | ") : '```' + JSON.stringify(args) + '```'
             await dummy.run(message, args)
         }
     }
